update DApp dashboard UX
